add routing to support both live dashboard and dump views
