+++ src/main/kotlin/App.kt
@@ -1,58 +1,35 @@
-import kotlinx.browser.document
-import kotlinx.html.InputType
-import kotlinx.html.id
-import kotlinx.html.js.onClickFunction
-import org.w3c.dom.HTMLInputElement
+import kotlinx.css.span
 import react.RBuilder
 import react.RComponent
 import react.RProps
 import react.RState
-import react.dom.button
-import react.dom.defaultValue
-import react.dom.h1
-import react.dom.input
-import react.dom.p
-import react.setState
+import react.dom.div
+import react.dom.span
+import react.router.dom.hashRouter
+import react.router.dom.route
+import react.router.dom.routeLink
+import react.router.dom.switch
 
-class App : RComponent<RProps, AppState>() {
-    init {
-        state.jvbUrl = null
-    }
+class App : RComponent<RProps, RState>() {
     override fun RBuilder.render() {
-        h1 {
-            +"JVB dashboard"
-        }
-        p {
-            key = "jvbUrl"
-            input(type = InputType.text) {
-                attrs {
-                    id = "jvb-url"
-                    defaultValue = "127.0.0.1:4443"
-                }
-            }
-            button {
-                attrs {
-                    text("Set JVB")
-                    onClickFunction = { _ ->
-                        val inputUrl = (document.getElementById("jvb-url") as HTMLInputElement).value
-                        console.log("setting jvb url to http://$inputUrl/debug")
-                        setState {
-                            jvbUrl = "http://$inputUrl/debug"
+        hashRouter {
+            switch {
+                route("/", exact = true) {
+                    div {
+                        routeLink("/dump") {
+                            +"Dump viewer"
+                        }
+                        span {
+                            +" | "
+                        }
+                        routeLink("/live") {
+                            +"Live dashboard"
                         }
                     }
                 }
-            }
-        }
-        if (!state.jvbUrl.isNullOrBlank()) {
-            child(Jvb::class) {
-                attrs {
-                    url = state.jvbUrl!!
-                }
+                route("/dump", Dump::class, exact = true)
+                route("/live", LiveDashboard::class, exact = true)
             }
         }
     }
-}
-
-external interface AppState : RState {
-    var jvbUrl: String?
 }
